+++ tests/assets/function8.ts
@@ -1,0 +1,3 @@
+export function foo(canManage = true) {
+  return !canManage;
+}